Fix failing test, revert changes to schemas which defines int type

+++ +types/+hdmf_common/ElementIdentifiers.m
@@ -16,7 +16,7 @@ methods
         %  elementIdentifiers = types.hdmf_common.ELEMENTIDENTIFIERS(Name, Value) creates a ElementIdentifiers object where one or more property values are specified using name-value pairs.
         %
         % Input Arguments (Name-Value Arguments):
-        %  - data (int32) - Data property for dataset class (ElementIdentifiers)
+        %  - data (int8) - Data property for dataset class (ElementIdentifiers)
         %
         % Output Arguments:
         %  - elementIdentifiers (types.hdmf_common.ElementIdentifiers) - A ElementIdentifiers object
@@ -39,7 +39,7 @@ methods
     %% VALIDATORS
     
     function val = validate_data(obj, val)
-        val = types.util.checkDtype('data', 'int32', val);
+        val = types.util.checkDtype('data', 'int8', val);
         types.util.validateShape('data', {[Inf]}, val)
     end
     %% EXPORT
